+++ test/reindexer_wrapper_test.go
@@ -184,7 +184,7 @@ func (dbw *ReindexerWrapper) execQueryCtx(ctx context.Context, qt *queryTest) *r
 func (dbw *ReindexerWrapper) setSlaveConfig(slaveDb *ReindexerWrapper) {
 	err := dbw.Upsert(reindexer.ConfigNamespaceName, reindexer.DBConfigItem{
 		Type:        "replication",
-		Replication: &reindexer.DBReplicationConfig{Namespaces: []string{}, Role: "master", ClusteID: 1, ForceSyncOnLogicError: true, ForceSyncOnWrongDataHash: true},
+		Replication: &reindexer.DBReplicationConfig{Namespaces: []string{}, Role: "master", ClusterID: 1, ForceSyncOnLogicError: true, ForceSyncOnWrongDataHash: true},
 	})
 	if err != nil {
 		panic(err)
@@ -192,7 +192,7 @@ func (dbw *ReindexerWrapper) setSlaveConfig(slaveDb *ReindexerWrapper) {
 
 	err = slaveDb.Upsert(reindexer.ConfigNamespaceName, reindexer.DBConfigItem{
 		Type:        "replication",
-		Replication: &reindexer.DBReplicationConfig{Namespaces: []string{}, Role: "slave", MasterDSN: *dsn, ClusteID: 1, ForceSyncOnLogicError: true, ForceSyncOnWrongDataHash: true},
+		Replication: &reindexer.DBReplicationConfig{Namespaces: []string{}, Role: "slave", MasterDSN: *dsn, ClusterID: 1, ForceSyncOnLogicError: true, ForceSyncOnWrongDataHash: true},
 	})
 	if err != nil {
 		panic(err)